# docs/swagger.yaml
definitions:
...
  disgm.Role:
    properties:
      color:
        description: Color of the role
        type: integer
      hoist:
        description: Whether the role is hoisted in the user list
        type: boolean
      id:
        description: Define fields for Role structure based on your needs
        type: string
      managed:
        description: Whether the role is managed by an application
        type: boolean
      mentionable:
        description: Whether the role is mentionable
        type: boolean
      name:
        description: Name of the role
        type: string
      permissions:
        description: Permissions for the role
        type: string
      position:
        description: Position of the role
        type: integer
    type: object
...
  models.AvatarDecorationData:
    properties:
      decoration:
        description: Example field for decoration
        type: string
    type: object
...
  models.Role:
    properties:
      color:
        description: Color of the role
        type: integer
      hoist:
        description: Whether the role is hoisted in the user list
        type: boolean
      id:
        description: Define fields for Role structure based on your needs
        type: string
      managed:
        description: Whether the role is managed by an application
        type: boolean
      mentionable:
        description: Whether the role is mentionable
        type: boolean
      name:
        description: Name of the role
        type: string
      permissions:
        description: Permissions for the role
        type: string
      position:
        description: Position of the role
        type: integer
    type: object
  models.RoleParams:
    properties:
      color:
        description: RGB color value
        type: integer
      hoist:
        description: Whether the role should be displayed separately in the sidebar
        type: boolean
      icon:
        description: The role's icon image (if the guild has the ROLE_ICONS feature)
        type: string
      mentionable:
        description: Whether the role should be mentionable
        type: boolean
      name:
        description: Name of the role, max 100 characters
        type: string
      permissions:
        description: Bitwise value of the enabled/disabled permissions
        type: string
      unicode_emoji:
        description: The role's unicode emoji as a standard emoji (if the guild has
          the ROLE_ICONS feature)
        type: string
    type: object
...
  models.User:
    properties:
      accent_color:
        description: Optional banner color as an integer
        type: integer
      avatar:
        description: Optional avatar hash of the user
        type: string
      avatar_decoration_data:
        allOf:
        - $ref: '#/definitions/models.AvatarDecorationData'
        description: Optional avatar decoration data
      banner:
        description: Optional banner hash of the user
        type: string
      bot:
        description: Optional flag indicating if the user is a bot
        type: boolean
      discriminator:
        description: User's Discord tag (four-digit identifier)
        type: string
      email:
        description: Optional user's email
        type: string
      flags:
        description: Optional flags on the user's account
        type: integer
      global_name:
        description: Optional display name (for bots, the application name)
        type: string
      id:
        description: Snowflake ID of the user
        type: string
      locale:
        description: Optional user's chosen language
        type: string
      mfa_enabled:
        description: Optional flag indicating if 2FA is enabled
        type: boolean
      premium_type:
        description: Optional Nitro subscription type
        type: integer
      public_flags:
        description: Optional public flags on the user's account
        type: integer
      system:
        description: Optional flag indicating if the user is a system user
        type: boolean
      username:
        description: Username of the user (not unique)
        type: string
      verified:
        description: Optional flag indicating if the email is verified
        type: boolean
    type: object
...
host: localhost:90
info:
  contact: {}
  description: API for managing Discord guilds using DiscordGo and Fiber.
  title: Discord Guild Management API
  version: "1.0"
paths:
...
  /api/guild/channels/{channelid}/messages/{messageid}/reactions:
    delete:
      description: Remove all reactions from a specific message in a channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete All Message Reactions
      tags:
      - Reactions
  /api/guild/channels/{channelid}/messages/{messageid}/reactions/{emojiid}:
    delete:
      description: Remove a specific emoji reaction from a message in a channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      - description: Emoji ID
        in: path
        name: emojiid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete Message Reaction Emoji
      tags:
      - Reactions
    get:
      description: Retrieve all reactions from a specific message in a channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      - description: Emoji ID
        in: path
        name: emojiid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.User'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Message Reactions
      tags:
      - Reactions
    put:
      description: Add a reaction to a specific message in a channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      - description: Emoji ID
        in: path
        name: emojiid
        required: true
        type: string
      responses:
        "201":
          description: Created
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create Message Reaction
      tags:
      - Reactions
  /api/guild/channels/{channelid}/messages/{messageid}/reactions/{emojiid}/{userid}:
    delete:
      description: Delete a user's reaction from a specific message in a channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      - description: Emoji ID
        in: path
        name: emojiid
        required: true
        type: string
      - description: User ID
        in: path
        name: userid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete Message Reaction
      tags:
      - Reactions
...
  /api/guild/roles:
    get:
      description: Retrieve all roles of a specific guild using the guild ID.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/disgm.Role'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all roles in a guild
      tags:
      - Roles
    patch:
      description: Reorder the roles in a guild based on the provided positions.
      parameters:
      - description: New role positions
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Role'
          type: array
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update role positions in a guild
      tags:
      - Roles
    post:
      description: Create a new role in a guild using the provided role parameters.
      parameters:
      - description: Role parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RoleParams'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Role'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create a new role in a guild
      tags:
      - Roles
  /api/guild/roles/{roleid}:
    delete:
      description: Delete a specific role from a guild using its role ID.
      parameters:
      - description: ID of the role to delete
        in: path
        name: roleid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete a role from a guild
      tags:
      - Roles
    get:
      description: Retrieve a specific role from a guild by its role ID.
      parameters:
      - description: ID of the role to retrieve
        in: path
        name: roleid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get a specific role in a guild
      tags:
      - Roles
    patch:
      description: Update a specific role in a guild using the provided role data.
      parameters:
      - description: ID of the role to update
        in: path
        name: roleid
        required: true
        type: string
      - description: Updated role parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RoleParams'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update a specific role in a guild
      tags:
      - Roles
...
swagger: "2.0"